modified the structure of solution the interfaces have been moved to the CarRental.Application.Interfaces folder, and folders have been created to contain DTO entities in CarRental.Application.Contracts
