--- CarRental.Application.Contracts/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace CarRental.Application.Contracts;
-
-public class Class1
-{
-
-}
